+++ app/assets/javascripts/actions/actionTypes.js
@@ -9,5 +9,7 @@ export default keyMirror({
   RECEIVE_STORIES: null,
   RECEIVE_PAST_ITERATIONS: null,
   TOGGLE_STORY: null,
-  EDIT_STORY: null
+  EDIT_STORY: null,
+  UPDATE_STORY: null,
+  UPDATE_STORY_SUCCESS: null
 });