+++ app/assets/javascripts/actions/story.js
@@ -10,6 +10,24 @@ export const toggleStory = (id) => ({
   id
 });
 
+export const updateStorySuccess = (story) => ({
+  type: actionTypes.UPDATE_STORY_SUCCESS,
+  story
+});
+
+export const updateStory = (story, projectId) => {
+  return (dispatch, getState, { Story }) => {
+    if (story._editing._isDirty) {
+      return Story.update(story._editing, projectId)
+        .then(( story ) => {
+          dispatch(updateStorySuccess(story));
+          dispatch(toggleStory(story.id));
+        });
+    }
+    return dispatch(toggleStory(story.id));
+  };
+};
+
 export const editStory = (id, newAttributes) => ({
   type: actionTypes.EDIT_STORY,
   id,