+++ app/assets/javascripts/components/story/ExpandedStory/ExpandedStoryControls.js
@@ -1,11 +1,21 @@
 import React from 'react';
 
 const ExpandedStoryControls = (props) => {
-  const { onCancel } = props;
+  const { onCancel, onSave } = props;
 
   return (
     <div className="form-group Story__controls">
-      <input className="cancel" onClick={onCancel} type="button" value={I18n.translate('cancel')} />
+      <input className="save"
+        onClick={onSave}
+        type="button"
+        value={I18n.t('save')}
+      />
+
+      <input className="cancel"
+        onClick={onCancel}
+        type="button"
+        value={I18n.t('cancel')}
+      />
     </div>
   );
 };