+++ app/assets/javascripts/models/beta/projectBoard.js
@@ -1,14 +1,15 @@
-import axios from 'axios';
-import camelCase from 'camelcase-object-deep';
+import httpService from '../../services/httpService';
+import changeCase from'change-object-case';
 
 export function get(projectId) {
-  return axios
+  return httpService
     .get(`/beta/project_boards/${projectId}`)
-    .then(({ data }) => camelCase(data, {deep: true}))
+    .then(({ data }) => changeCase.camelKeys(data, {recursive: true, arrayRecursive: true}))
     .then(( projectBoard ) => ({
         ...projectBoard,
         stories: projectBoard.stories.map((story) => ({
           ...story,
+          estimate: story.estimate || '',
           collapsed: true
         }))
       }));